backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0.post1
websockets==12.0
openai==1.14.0
python-dotenv==1.0.1
pydantic==2.6.1
numpy>=2.2.0,<3.0.0
tensorflow==1.15.0
openai-agents[voice]==0.0.6
python-multipart==0.0.6
deepgram-sdk==3.10.1
pyaudio==0.2.13 